[Refactor component] 컴포넌트 분할 수술 (#22)
* refactor 컴포넌트 구분 분할 80%

* feat 로그인 후 이동
