+++ src/components/layouts/AppHeader.tsx
@@ -200,9 +200,6 @@ const AppHeader = () => {
     <App scaleheight={scaleHeight.toString()}>
       <div className="header__container">
         <NavBtn>
-          {/* <Button type="text" size="large" onClick={handleShowSider}>
-            {showSider ? <CloseOutlined /> : <MenuOutlined />}
-          </Button> */}
           <div>
             <Link href="/">Kormelon Devlog</Link>
           </div>
@@ -243,7 +240,7 @@ const AppHeader = () => {
           {user.id ? (
             <>
               <Button type="link" size="middle">
-                <Link href="/upload">Upload</Link>
+                <Link href="/write">write</Link>
               </Button>
               <Button type="link" size="middle" onClick={handleLogOut}>
                 Logout
