+++ src/components/layouts/AppLoading.tsx
@@ -1,9 +1,15 @@
 import { Spin } from 'antd';
 import { Content } from 'antd/lib/layout/layout';
 import React from 'react';
-import AppContents from './AppContents';
 import styled from '@emotion/styled';
 
+const ContentBlock = styled.div`
+  width: 100%;
+  min-height: 100vh;
+  padding: 1.25em;
+  max-width: 1000px;
+  margin: 105px auto 0 auto;
+`;
 const Loading = styled.div`
   position: fixed;
   top: 50%;
@@ -22,14 +28,14 @@ const AppLoading = ({ text, scroll }: any) => {
   }
 
   return (
-    <AppContents>
+    <ContentBlock>
       <Content>
         <Loading>
           <Spin size="large"></Spin>
           {text || <div>글 불러오는 중</div>}
         </Loading>
       </Content>
-    </AppContents>
+    </ContentBlock>
   );
 };
