+++ src/interfaces/categories.ts
@@ -1,0 +1,5 @@
+export interface Categories {
+  _id?: string;
+  category: string;
+  post: any[];
+}
