+++ src/pages/404/index.tsx
@@ -1,8 +1,8 @@
 import React from 'react';
-import AppEmpty from '../../components/layouts/AppEmpty';
+import ContentEmpty from '../../components/home/common/ContentEmpty';
 
 const index = () => {
-  return <AppEmpty text={'아직 준비 중인 페이지 입니다. 조금만 기다려주세요.'}></AppEmpty>;
+  return <ContentEmpty text={'아직 준비 중인 페이지 입니다. 조금만 기다려주세요.'}></ContentEmpty>;
 };
 
 export default index;
