+++ src/pages/write.tsx
@@ -1,0 +1,3 @@
+import Write from '../components/write';
+
+export default Write;
